Faster check for conflicts before running every command.
This patch adds a hasConflicts(String ns):boolean method to
StagingDatabase as a faster check for ConflictInterceptor
to verify if the command cannot be run when conflicts exist.

Previously, ConflictInterceptor was loading the full list
of conflicts to check for its size, which can be really slow
if the number of conflicts is large.