remove touchedId lists in Local/HttpRemoteRepo.
these in-memory lists are there only to clean up after a possible exception,
but they can become huge and consume too much heap. Getting rid of this clean
up eliminates the overhead. If an error occurred that leaves the repository
with too much garbage we could still make a clean clone though.