Fix bug on sparse cloning missing feature-only metadata ids and other small improvements.
BinaryPackedChanges was exporting the RevFeatureType metadata only from
trees, but the object model allows for specific features inside a RevTree
not to share the tree node's metadata id.

Removed the CAP instance variable and changesSent method variable
since changesSent was not being incremented and hence it was dead code.

Made it so a byte[] does not need to be cloned for each object written.