Fix DiffCount to return the correct number of affected trees.
DiffCount was computing the correct number of affected trees
of a diff only if a tree was fully added or removed. The test
expectations were also wrong.

This patch fixes DiffCountConsumer to increment the tree counter
when there are deeper diffs to check too, fixes the unit and
functional test expectations that were wrong, and removes the
unnecessary DiffCount.setReportTrees(boolean) setting.