Huge performance improvement to compute the number of changes with path filtering.
This patch introduces two new methods to the DiffTreeVisitor.Consumer interface:
endTree(Node left, Node right):void and
endBucket(int bucketIndex, int bucketDepth, Bucket left, Bucket right)
that are called after the tree(...):boolean and bucket(...):bookean respectively,
regardless of their return value.

This allows consumers to keep track of more info like what the current tree
path is at any given event.

With that in mind, DiffCountConsumer is updated to use a filtering decorator
when path filters are given, and hence get rid of the separate code path it
was using for that case, that forced a full traversal of a DiffTreeIterator.

This results in an impressive performance improvement as now DiffCountConsumer
can skip whole trees and buckets and address only the ones that can possibly
match one of the path filters.

In a repo of ~2M osm features with ~1M changes between the diff'ed commits,
and filtering with 10 path filters, we're talking of an avg execution time
of ~5 milliseconds as opposed to ~3000 milliseconds.