Remove the -XX:UseCompressedOops JVM argument in the generated CLI scripts.
The argument was there regardless of compressedoops being the default
on newer 64-bit JVM's because the BDB JE documentation mentioned its
needed anyways for it to correcly estimate cache usage size.
The current version of BDB JE is able to detect the setting even if
not explicitly set:

java -cp ~/.m2/repository/com/sleepycat/je/5.0.104/je-5.0.104.jar com.sleepycat.je.dbi.CompressedOopsDetector
isEnabled(): true

Hence we remove the JVM argument because it causes issues when geogit
is run with 32-bit JVM's and is not needed by BDBJE at least on Oracle JVMs.